crew-app/requirements.txt
crewai
crewai-tools
litellm
boto3
pydantic
pydantic-core
python-dotenv
langchain-community
litellm[proxy]
gradio
requests
librosa
numpy
scipy
soundfile